# downstream/src/test/resources/com/example/connect/downstream/OpenApiTests.generates open-api docs.approved.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Downstream API",
    "version" : "v0.0.1"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Local server"
  } ],
  "paths" : {
    "/movies/{title}/reviews" : {
      "post" : {
        "tags" : [ "review-movie-controller" ],
        "summary" : "Review movie",
        "operationId" : "reviewMovie",
        "parameters" : [ {
          "name" : "title",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "X-Stub-Response",
          "in" : "header",
          "schema" : {
            "type" : "string",
            "enum" : [ "review-submitted" ]
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Created"
          }
        }
      }
    },
    "/movies" : {
      "get" : {
        "tags" : [ "list-movies-controller" ],
        "summary" : "List movies",
        "operationId" : "listMovies",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          }
        }
      }
    },
    "/movies/{title}" : {
      "get" : {
        "tags" : [ "get-movie-details-controller" ],
        "summary" : "Get movie details",
        "operationId" : "movieDetails",
        "parameters" : [ {
          "name" : "title",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "Not Found"
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : { }
}